Specify PGID and PUID to our services
This allows downloaded media to end up with the same user and group as
the one who's running docker. Since these are mounted as volumes on the
host machine, it's a little bit cleaner to have the correct permissions
there.

It requires to launch `docker-compose` as follows:

```shell
PGID="$(id -g)" PUID="$(id -u)" docker-compose up -d
```