Make the environment more portable
Volumes are now part of the current directory. This way, the whole
setup can be started from absolutely anywhere, without further
configuration (no need to create a `media` user for instance).

The conf. files are defaults, one will require to visit the
configuration interface of every service to get a working setup.