Include the reverse proxy in the compose setup
The NGinx reverse-proxy mapping to all our sub services is now part of
the `docker-compose.yaml` definitions. Spawning a plex server on an
brand new machine will now be as simple as `docker-compose up -d`.

A user already using a reverse proxy on its host would just keep the
setup as is (incoming traffic on port 8080) and have a `/media-server`
(for instance) sub route redirect onto port 8080. The `url_base`
options in Couchpotato's and Sonarr's conf may need to be altered too.

A user willing to run the setup directly on the host's root URL path
would rather enable HTTPS in the NGinx configuration (with HTTP
redirection) by uncommenting the appropriate lines in
`conf/nignx/nginx.conf` (and also deleting some). Updating port
mappings from `8080:80` to `443:443` will also be required.