Add plex media server to the compose setup
Plex has just been added to our Docker-compose setup, the port 5353
(local Network discovery) has been shutdown. If running Plex on
premises, at home, it might be interesting to re-enable it though.

There's one thing to change though: it would be great to have the
gateway forward traffic for plex on all ports. This would allow us to
put our SSL certificate in NGinx and use it for SSL termination as
well.

This way, our setup would be similar to a local VPN on which traffic is
broadcast in clear (which is not a problem at all since it doesn't
leave the machine) but with an SSL gateway between the public (the
host network here) and the VPN.