Re-enable Plex port :-( (and add Jackett)
We need that port opened for remote access...

Oh and Jackett is a proxy to many torrent trackers, both public and
private. It enables access for Sonnar to The Pirate Bay's multiple
proxies, as well as kat.cr (which is back). It think we should enable a
few good public torrent trackers in our default config, it would make
the setup process even smoother.